requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy
asyncpg
supabase==1.0.3
pydantic
python-dotenv==1.0.0

# News Analysis Dependencies
exa-py==1.8.8
google-generativeai==0.3.1
deep-translator==1.11.4
langdetect==1.0.9
cachetools==5.3.2
aiohttp==3.9.1
python-multipart==0.0.6

